Model/predict_funciton/requirements.txt
google_cloud_storage==2.4.0
pandas
numpy
scikit-learn
tensorflow==2.9.1
tensorflow_hub
flask 
functions_framework
Pillow==9.1.0